backend/pyproject.toml
[tool.poetry]
name = "Zensai todo backend"
version = "0.1.0"
description = "Django backend for Zensai todo app"
authors = ["Charles Apochi <charlesapochi@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
graphene-django = "^3.2.2"
drf-spectacular = "^0.27.2"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
django-cors-headers = "^4.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
